refactor : 도메인 - 인프라 공용 언어
