+++ src/main/java/com/icestudyroom_email/domain/contract/ranking/PersonalNotificationPayload.java
@@ -1,4 +1,4 @@
-package com.icestudyroom_email.domain.rankingContract;
+package com.icestudyroom_email.domain.contract.ranking;
 
 public record PersonalNotificationPayload(
         String type
