+++ src/main/java/com/icestudyroom_email/domain/contract/ranking/RankingListEvent.java
@@ -1,4 +1,4 @@
-package com.icestudyroom_email.domain.rankingContract;
+package com.icestudyroom_email.domain.contract.ranking;
 
 import java.util.List;
