+++ src/main/java/com/icestudyroom_email/domain/contract/ranking/RankingPeriod.java
@@ -1,4 +1,4 @@
-package com.icestudyroom_email.domain.rankingContract;
+package com.icestudyroom_email.domain.contract.ranking;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
